requirements.txt
flask~=3.0
python-dotenv~=1.1
google-genai~=1.29
